Add swagger in order service and add create order method

+++ ms_proj_01_order-management-platform/order-service/src/main/java/com/mh/crj/model/OrderRequestDto.java
@@ -5,8 +5,8 @@ import lombok.Data;
 @Data
 public class OrderRequestDto {
 
-	private Long userId;
-    private Long productId;
+	private Integer userId;
+    private Integer productId;
     private Integer quantity;
     private Double totalAmount;
